delete tips page
